requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cloudinary==1.28.0
confusable-homoglyphs==3.2.0
Django==3.2.9
django-bootstrap4==21.2
django-braces==1.15.0
django-model-utils==4.2.0
django-registration==3.1.2
django-registration-redux==2.9
django-star-ratings==0.9.2
Pillow==9.0.0
psycopg2==2.9.3
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.8
